requirements.txt
absl-py==1.4.0
accelerate==0.26.1
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.25.0
anyio==4.9.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.37.3
botocore==1.37.3
Brotli @ file:///croot/brotli-split_1736182456865/work
bs4==0.0.2
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi==1.17.1
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cloudpickle==3.0.0
cryptography==45.0.3
datasets==3.6.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
filelock @ file:///croot/filelock_1744281381737/work
frozenlist==1.6.0
fsspec==2025.3.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
llvmlite==0.43.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.4.4
multiprocess==0.70.16
networkx @ file:///croot/networkx_1717597493534/work
numba==0.60.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1744613067434/work
platformdirs==4.3.8
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
retrying==1.3.4
s3fs==2025.5.1
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.1
shap==0.47.2
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-ml-python==1.8.5
snowflake-snowpark-python==1.32.0
snowflake.core==1.5.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
suds==1.2.0
sympy @ file:///croot/sympy_1738108488918/work
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
tomlkit==0.13.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.37.2
triton==2.1.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
wrapt==1.17.2
xgboost==2.1.4
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
